@@ -1,6 +1,6 @@
 {
   "name": "svelte",
-  "version": "3.42.6",
+  "version": "3.43.1",
   "description": "Cybernetically enhanced web apps",
   "module": "index.mjs",
   "main": "index",
@@ -22,6 +22,10 @@
   "exports": {
     "./package.json": "./package.json",
     ".": {
+      "node": {
+        "import": "./ssr.mjs",
+        "require": "./ssr.js"
+      },
       "import": "./index.mjs",
       "require": "./index.js"
     },
@@ -110,17 +114,17 @@
     "@sveltejs/eslint-config": "github:sveltejs/eslint-config#v5.7.0",
     "@types/mocha": "^7.0.0",
     "@types/node": "^8.10.53",
-    "@typescript-eslint/eslint-plugin": "^4.9.0",
-    "@typescript-eslint/parser": "^4.9.0",
+    "@typescript-eslint/eslint-plugin": "^4.31.2",
+    "@typescript-eslint/parser": "^4.31.2",
     "acorn": "^8.4.1",
     "agadoo": "^1.1.0",
     "c8": "^5.0.1",
     "code-red": "^0.2.2",
     "codecov": "^3.5.0",
     "css-tree": "^1.1.2",
-    "eslint": "^7.15.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-svelte3": "^2.7.3",
+    "eslint": "^7.32.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-svelte3": "^3.2.1",
     "estree-walker": "^3.0.0",
     "is-reference": "^3.0.0",
     "jsdom": "^15.2.1",
