@@ -1,4 +1,4 @@
-const sh = require('shelljs');
+import sh from 'shelljs';
 
 sh.rm('-rf', 'static/workers');
 sh.cp('-r', 'node_modules/@sveltejs/svelte-repl/workers', 'static');
