@@ -1,9 +1,11 @@
-const sh = require('shelljs');
-const fs = require('fs');
-const path = require('path');
+import sh from 'shelljs';
+import fs from 'fs';
+import path, { dirname }  from 'path';
+import { fileURLToPath } from 'url';
 
 const force = process.env.FORCE_UPDATE === 'true';
 
+const __dirname = dirname(fileURLToPath(import.meta.url));
 sh.cd(path.join(__dirname, '..'));
 
 const outputFile = 'static/svelte-app.json';
