@@ -20,6 +20,9 @@ if (!dev && !process.env.MAPBOX_ACCESS_TOKEN) {
 const onwarn = (warning, onwarn) => (warning.code === 'CIRCULAR_DEPENDENCY' && /[/\\]@sapper[/\\]/.test(warning.message)) || onwarn(warning);
 const dedupe = importee => importee === 'svelte' || importee.startsWith('svelte/');
 
+const serverOutput = config.server.output();
+serverOutput.format = 'esm';
+
 export default {
 	client: {
 		input: config.client.input(),
@@ -71,7 +74,7 @@ export default {
 
 	server: {
 		input: config.server.input(),
-		output: config.server.output(),
+		output: serverOutput,
 		plugins: [
 			replace({
 				'process.browser': false,
