@@ -1,5 +1,5 @@
 <script>
-	let clicked;
+	export let clicked;
 </script>
 
 <button on:click='{() => clicked = true}'>click me</button>
